+++ src/main/java/com/soundcloud/api/CloudAPI.java
@@ -22,7 +22,7 @@ public interface CloudAPI {
     // other constants
     String REALM              = "SoundCloud";
     String OAUTH_SCHEME       = "oauth";
-    String VERSION            = "1.0.0";
+    String VERSION            = "1.0.1";
     String USER_AGENT         = "SoundCloud Java Wrapper ("+VERSION+")";
 
     /**
@@ -37,10 +37,24 @@ public interface CloudAPI {
      */
     Token login(String username, String password) throws IOException;
 
+    /**
+     * Request a token using <a href="http://tools.ietf.org/html/draft-ietf-oauth-v2-10#section-4.1.2">
+     * Resource Owner Password Credentials</a>.
+     *
+     * @param username SoundCloud username
+     * @param password SoundCloud password
+     * @param scope    the desired scope
+     * @return a valid token
+     * @throws com.soundcloud.api.CloudAPI.InvalidTokenException
+     *                     invalid token
+     * @throws IOException In case of network/server errors
+     */
+    Token login(String username, String password, String scope) throws IOException;
+
 
     /**
      * Request a token using <a href="http://tools.ietf.org/html/draft-ietf-oauth-v2-10#section-4.1.1">
-     * Authorization Code</a>.
+     * Authorization Code</a>, requesting a default scope.
      *
      * @param code the authorization code
      * @return a valid token
@@ -48,6 +62,18 @@ public interface CloudAPI {
      * @throws IOException In case of network/server errors
      */
     Token authorizationCode(String code) throws IOException;
+
+    /**
+     * Request a token using <a href="http://tools.ietf.org/html/draft-ietf-oauth-v2-10#section-4.1.1">
+     * Authorization Code</a> with a specified scope.
+     *
+     * @param code the authorization code
+     * @param scope the desired scope
+     * @return a valid token
+     * @throws com.soundcloud.api.CloudAPI.InvalidTokenException invalid token
+     * @throws IOException In case of network/server errors
+     */
+    Token authorizationCode(String code, String scope) throws IOException;
 
     /**
      * Request a "signup" token using <a href="http://tools.ietf.org/html/draft-ietf-oauth-v2-15#section-4.4">
@@ -169,11 +195,19 @@ public interface CloudAPI {
      * <li><code>error</code> in case of failure, this contains an error code (most likely
      * <code>access_denied</code>).
      * </ul>
-     * @param  options auth endpoint to use (leave out for default)
+     * @param  options auth endpoint to use (leave out for default), requested scope (leave out for default)
      * @return the URI to open in a browser/WebView etc.
      * @see CloudAPI#authorizationCode(String)
      */
     URI authorizationCodeUrl(String... options);
+
+    /**
+     * Changes the default content type sent in the "Accept" header.
+     * If you don't set this it defaults to "application/json".
+     *
+     * @param contentType the request mime type.
+     */
+    void setDefaultContentType(String contentType);
 
     /**
      * Interested in changes to the current token.