+++ src/test/java/com/soundcloud/api/CloudAPIIntegrationTest.java
@@ -1,6 +1,7 @@
 package com.soundcloud.api;
 
 import static org.hamcrest.CoreMatchers.*;
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.junit.Assert.assertThat;
 
 import org.apache.http.HttpResponse;
@@ -9,9 +10,11 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.ByteBuffer;
 
 public class CloudAPIIntegrationTest implements Params.Track, Endpoints {
     // http://sandbox-soundcloud.com/you/apps/java-api-wrapper-test-app
@@ -36,12 +39,19 @@ public class CloudAPIIntegrationTest implements Params.Track, Endpoints {
                 null,
                 null,
                 Env.SANDBOX);
+    }
 
-        api.login("api-testing", "testing");
+    private Token login() throws IOException {
+        return login(null);
+    }
+
+    private Token login(String scope) throws IOException {
+        return api.login("api-testing", "testing", scope);
     }
 
     @Test
     public void shouldUploadASimpleAudioFile() throws Exception {
+        login();
         HttpResponse resp = api.post(Request.to(TRACKS).with(
                   TITLE, "Hello Android",
                   POST_TO_EMPTY, "")
@@ -51,13 +61,34 @@ public class CloudAPIIntegrationTest implements Params.Track, Endpoints {
         assertThat(status, is(201));
     }
 
+    @Test
+    public void shouldUploadASimpleAudioFileBytes() throws Exception {
+        login();
+
+        File f = new File(getClass().getResource("hello.aiff").getFile());
+        ByteBuffer bb = ByteBuffer.allocate((int) f.length());
+        FileInputStream fis = new FileInputStream(f);
+        for (;;) if (fis.getChannel().read(bb) <= 0) break;
+
+        HttpResponse resp = api.post(Request.to(TRACKS).with(
+                  TITLE, "Hello Android",
+                  POST_TO_EMPTY, "")
+                .withFile(ASSET_DATA, bb));
+
+        int status = resp.getStatusLine().getStatusCode();
+        assertThat(status, is(201));
+    }
+
+
     @Test(expected = IOException.class)
     public void shouldNotGetASignupTokenWhenInofficialApp() throws Exception {
+        login();
         api.clientCredentials();
     }
 
     @Test
     public void shouldReturn401WithInvalidToken() throws Exception {
+        login();
         api.setToken(new Token("invalid", "invalid"));
         HttpResponse resp = api.get(Request.to(Endpoints.MY_DETAILS));
         assertThat(resp.getStatusLine().getStatusCode(), is(401));
@@ -65,6 +96,8 @@ public class CloudAPIIntegrationTest implements Params.Track, Endpoints {
 
     @Test
     public void shouldRefreshAutomaticallyWhenTokenExpired() throws Exception {
+        login();
+
         HttpResponse resp = api.get(Request.to(Endpoints.MY_DETAILS));
         assertThat(resp.getStatusLine().getStatusCode(), is(200));
 
@@ -80,19 +113,61 @@ public class CloudAPIIntegrationTest implements Params.Track, Endpoints {
 
     @Test
     public void shouldResolveUrls() throws Exception {
+        login();
+
         long id = api.resolve("http://sandbox-soundcloud.com/api-testing");
         assertThat(id, is(1862213L));
     }
 
     @Test
     public void readMyDetails() throws Exception {
+        login();
+
         HttpResponse resp = api.get(Request.to(Endpoints.MY_DETAILS));
         assertThat(resp.getStatusLine().getStatusCode(), is(200));
+
+        assertThat(
+                resp.getFirstHeader("Content-Type").getValue(),
+                containsString("application/json"));
 
         JSONObject me = Http.getJSON(resp);
 
         assertThat(me.getString("username"), equalTo("api-testing"));
         // writeResponse(resp, "me.json");
+    }
+
+    @Test
+    public void shouldLoginWithNonExpiringScope() throws Exception {
+        Token token = login(Token.SCOPE_NON_EXPIRING);
+        assertThat(token.scoped(Token.SCOPE_NON_EXPIRING), is(true));
+        assertThat(token.refresh, is(nullValue()));
+        assertThat(token.getExpiresIn(), is(nullValue()));
+        assertThat(token.valid(), is(true));
+
+        // make sure we can issue a request with this token
+        HttpResponse resp = api.get(Request.to(Endpoints.MY_DETAILS));
+        assertThat(resp.getStatusLine().getStatusCode(), is(200));
+    }
+
+    @Test
+    public void shouldNotRefreshWithNonExpiringScope() throws Exception {
+        Token token = login(Token.SCOPE_NON_EXPIRING);
+        assertThat(token.scoped(Token.SCOPE_NON_EXPIRING), is(true));
+        assertThat(api.invalidateToken(), is(nullValue()));
+        HttpResponse resp = api.get(Request.to(Endpoints.MY_DETAILS));
+        assertThat(resp.getStatusLine().getStatusCode(), is(401));
+    }
+
+    @Test
+    public void shouldChangeContentType() throws Exception {
+        login();
+
+        api.setDefaultContentType("application/xml");
+        HttpResponse resp = api.get(Request.to(Endpoints.MY_DETAILS));
+
+        assertThat(
+                resp.getFirstHeader("Content-Type").getValue(),
+                containsString("application/xml"));
     }
 
     /*