+++ CHANGES.md
@@ -1,3 +1,13 @@
+## 1.0.1 2011-07-04
+
+  * Support for non-expiring scope
+  * Cancellation of uploads
+  * Added Request#withFile(String, byte[]), Request#withFile(String, ByteBuffer)
+  * Added Request#withEntity(HttpEntity) and Request#withContent(String, String)
+  * Added PutResource example
+  * Added setting to change the default content type for requests
+    (CloudAPI#setDefaultContentType(String))
+
 ## 1.0.0 2011-05-19
 
   * Initial release